Checks if the tag you're adding has a line break character at the end of it and deletes it
I was trying to debug my code for about an hour since it wasn't able to detect the new tag that I had added and set to a model using the tag editor. It turns out that an invisible character equal to string.char(13) was added to the end of the name, making it so that it was technically a different name than the one I was checking for in my script. So there should be some checking for this character at the end of the name of tags when they're being added to avoid this confusion.