Localization support (#71)
Now there is a `Localization.csv` that can have translations added.

I couldn't get `AutoLocalize` to work literally at all, it must be disabled in PluginGuis. So I had to go through the entire codebase and replace hardcoded strings with translation keys, and pass them to a new `tr()` function.

This doesn't actually add support for any new languages yet, but if anybody wants to help with that, it should be much easier now, since now the only tools needed are Google Sheets or LibreOffice Calc for editing CSV files.